+++ package.json
@@ -18,19 +18,19 @@
     "email": "plugins@grafana.com"
   },
   "devDependencies": {
-    "@babel/core": "^7.0.0",
-    "@types/d3": "^5.0.0",
-    "@types/d3-hexbin": "0.2.2",
-    "@types/jest": "^22.2.3",
-    "@types/lodash": "^4.14.116",
+    "@babel/core": "^7.2.2",
+    "@types/d3": "^5.0.1",
+    "@types/d3-hexbin": "0.2.3",
+    "@types/jest": "^23.3.12",
+    "@types/lodash": "^4.14.119",
     "babel-core": "^7.0.0-0",
     "babel-jest": "^23.4.2",
-    "csstype": "^2.5.6",
+    "csstype": "^2.6.0",
     "d3": "^5.7.0",
     "d3-hexbin": "^0.2.2",
     "grafana-sdk-mocks": "github:grafana/grafana-sdk-mocks",
     "grunt": "^1.0.3",
-    "grunt-contrib-clean": "^1.1.0",
+    "grunt-contrib-clean": "^2.0.0",
     "grunt-contrib-concat": "^1.0.1",
     "grunt-contrib-copy": "^1.0.0",
     "grunt-contrib-watch": "^1.1.0",
@@ -43,14 +43,13 @@
     "jest": "^23.5.0",
     "jquery": "^3.3.1",
     "load-grunt-tasks": "^4.0.0",
-    "moment": "^2.22.2",
-    "node-sass": "^4.9.3",
-    "plugin-typescript": "^8.0.0",
-    "regenerator-runtime": "^0.12.1",
-    "sass": "^1.13.1",
-    "ts-jest": "^23.1.4",
-    "tslint": "^5.11.0",
-    "typescript": "^2.9.2",
+    "moment": "^2.23.0",
+    "node-sass": "^4.11.0",
+    "regenerator-runtime": "^0.13.1",
+    "sass": "^1.16.0",
+    "ts-jest": "^23.10.5",
+    "tslint": "^5.12.1",
+    "typescript": "^3.2.2",
     "typescript-formatter": "^7.2.2"
   },
   "dependencies": {},
