+++ src/specs/timeseries.ts
